Giorno_6/requirements.txt
# Azure AI Dependencies
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-core>=1.29.0

# LangChain Dependencies (versioni aggiornate)
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Document Processing
PyPDF2>=3.0.0
pypdf>=3.17.0

# Vector Store
faiss-cpu>=1.7.4

# Scientific Computing
numpy>=1.24.0
scikit-learn>=1.3.0

# Environment Management
python-dotenv>=1.0.0

# Optional: For better PDF processing
pymupdf>=1.23.0
pdfplumber>=0.9.0